feat(host): Support multiple entrypoints
