+++ bin/host/src/eth/mod.rs
@@ -1,0 +1,10 @@
+//! Ethereum utilities for the host binary.
+
+mod blobs;
+pub use blobs::{
+    APIConfigResponse, APIGenesisResponse, OnlineBlobProvider, ReducedConfigData,
+    ReducedGenesisData,
+};
+
+mod precompiles;
+pub(crate) use precompiles::execute;
