+++ bin/host/src/kv/mod.rs
@@ -14,9 +14,6 @@ pub use disk::DiskKeyValueStore;
 mod split;
 pub use split::SplitKeyValueStore;
 
-mod local;
-pub use local::LocalKeyValueStore;
-
 /// A type alias for a shared key-value store.
 pub type SharedKeyValueStore = Arc<RwLock<dyn KeyValueStore + Send + Sync>>;
