+++ bin/host/src/server.rs
@@ -19,11 +19,12 @@ use tracing::{error, info};
 /// The [PreimageServer] is responsible for waiting for incoming preimage requests and
 /// serving them to the client.
 #[derive(Debug)]
-pub struct PreimageServer<P, H, KV>
+pub struct PreimageServer<P, H, KV, F>
 where
     P: PreimageOracleServer,
     H: HintReaderServer,
     KV: KeyValueStore + ?Sized,
+    F: Fetcher,
 {
     /// The oracle server.
     oracle_server: P,
@@ -33,14 +34,15 @@ where
     kv_store: Arc<RwLock<KV>>,
     /// The fetcher for fetching preimages from a remote source. If [None], the server will only
     /// serve preimages that are already in the key-value store.
-    fetcher: Option<Arc<RwLock<Fetcher<KV>>>>,
+    fetcher: Option<Arc<RwLock<F>>>,
 }
 
-impl<P, H, KV> PreimageServer<P, H, KV>
+impl<P, H, KV, F> PreimageServer<P, H, KV, F>
 where
     P: PreimageOracleServer + Send + Sync + 'static,
     H: HintReaderServer + Send + Sync + 'static,
     KV: KeyValueStore + Send + Sync + ?Sized + 'static,
+    F: Fetcher + Send + Sync + 'static,
 {
     /// Create a new [PreimageServer] with the given [PreimageOracleServer],
     /// [HintReaderServer], and [KeyValueStore]. Holds onto the file descriptors for the pipes
@@ -49,7 +51,7 @@ where
         oracle_server: P,
         hint_reader: H,
         kv_store: Arc<RwLock<KV>>,
-        fetcher: Option<Arc<RwLock<Fetcher<KV>>>>,
+        fetcher: Option<Arc<RwLock<F>>>,
     ) -> Self {
         Self { oracle_server, hint_reader, kv_store, fetcher }
     }
@@ -75,7 +77,7 @@ where
     /// client.
     async fn start_oracle_server(
         kv_store: Arc<RwLock<KV>>,
-        fetcher: Option<Arc<RwLock<Fetcher<KV>>>>,
+        fetcher: Option<Arc<RwLock<F>>>,
         oracle_server: P,
     ) -> Result<()> {
         #[inline(always)]
@@ -106,10 +108,7 @@ where
 
     /// Starts the hint router, which waits for incoming hints and routes them to the appropriate
     /// handler.
-    async fn start_hint_router(
-        hint_reader: H,
-        fetcher: Option<Arc<RwLock<Fetcher<KV>>>>,
-    ) -> Result<()> {
+    async fn start_hint_router(hint_reader: H, fetcher: Option<Arc<RwLock<F>>>) -> Result<()> {
         #[inline(always)]
         async fn do_loop<R, H>(router: &R, server: &H) -> Result<()>
         where
